prueba de nuevo token

+++ ToDoList/Views/ListView.swift
@@ -12,12 +12,23 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-            }.onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
-        }.listStyle(.plain)
+		ZStack{
+			if listViewModel.items.isEmpty {
+				Text("No items")
+			} else{
+				List {
+					ForEach(listViewModel.items) { item in
+						ListRowView(item: item)
+							.onTapGesture {
+								withAnimation(.linear){
+									listViewModel.updateItem(item: item)
+								}
+							}
+					}.onDelete(perform: listViewModel.deleteItem)
+						.onMove(perform: listViewModel.moveItem)
+				}.listStyle(.plain)
+			}
+		}
         .navigationTitle("Todo List 📝")
         .toolbar {
             ToolbarItem(placement: .navigationBarLeading,
